Rename tests and improve some
